requirements.txt
google-api-python-client==2.86.0
youtube-transcript-api==0.6.0
python-dotenv==1.0.0
